+++ drivers/luxtronik-2/device.js
@@ -13,6 +13,10 @@ const {
   getWritableParameter,
   validateTemperatureRawValue,
 } = require('../../lib/LuxtronikWriteSupport');
+const {
+  readCurrentPower,
+  readEnergyInputs,
+} = require('../../lib/LuxtronikEnergySupport');
 
 const COOLING_RELEASE_TEST_ORIGINAL_STORE_KEY = 'coolingReleaseTemperatureTestOriginalRaw';
 
@@ -100,6 +104,7 @@ class LuxtronikDevice extends Device {
     this.calulationsArray = null;
     this.previousCoolingDiagnosticState = null;
     this.coolingReleaseTemperatureTestOriginalRaw = this.getStoreValue(COOLING_RELEASE_TEST_ORIGINAL_STORE_KEY);
+    this.coolingReleaseOperationQueue = Promise.resolve();
 
     this.scan();
 
@@ -184,7 +189,7 @@ class LuxtronikDevice extends Device {
       }
       if (this.energyWater !== null) await this.setCapabilityValue('meter_power.water', this.energyWater / 10).catch(this.error);
       if (this.energyPool !== null) await this.setCapabilityValue('meter_power.pool', this.energyPool / 10).catch(this.error);
-      if (this.parametersArray !== null) await this.setCapabilityValue('meter_power.heat2', this.parametersArray[1059] / 10).catch(this.error);
+      if (Number.isFinite(this.parametersArray?.[1059])) await this.setCapabilityValue('meter_power.heat2', this.parametersArray[1059] / 10).catch(this.error);
 
       if (this.parametersArray !== null && this.calulationsArray !== null && this.energyTotal !== null) {
         this.log("Not null")
@@ -207,7 +212,7 @@ class LuxtronikDevice extends Device {
       // if (this.energyInputWater !== null) await this.setCapabilityValue('meter_power.water', this.energyWater / 100).catch(this.error);
       // if (this.energyInputPool !== null) await this.setCapabilityValue('meter_power.pool', this.energyPool / 100).catch(this.error);
 
-      if (this.energyCurrent !== null) await this.setCapabilityValue('measure_power.current', this.energyCurrent).catch(this.error);
+      if (Number.isFinite(this.energyCurrent)) await this.setCapabilityValue('measure_power.current', this.energyCurrent).catch(this.error);
 
       if (this.temperatureOutdoor !== null) await this.setCapabilityValue('measure_temperature.outdoor', this.temperatureOutdoor / 10).catch(this.error);
       if (this.temperatureHotGas !== null) await this.setCapabilityValue('measure_temperature.hotgas', this.temperatureHotGas / 10).catch(this.error);
@@ -580,14 +585,17 @@ class LuxtronikDevice extends Device {
     }
   }
 
-  async testWriteCoolingReleaseTemperatureNoop() {
+  testWriteCoolingReleaseTemperatureNoop() {
+    return this.queueCoolingReleaseOperation(() => this.testWriteCoolingReleaseTemperatureNoopQueued());
+  }
+
+  async testWriteCoolingReleaseTemperatureNoopQueued() {
     const parameter = VERIFIED_LUXTRONIK_PARAMETERS.COOLING_RELEASE_TEMPERATURE;
     const parameters = await this.readParameters();
     const currentRawValue = parameters[parameter.index];
     validateTemperatureRawValue(parameter, currentRawValue);
 
-    await this.setStoreValue(COOLING_RELEASE_TEST_ORIGINAL_STORE_KEY, currentRawValue);
-    this.coolingReleaseTemperatureTestOriginalRaw = currentRawValue;
+    await this.saveCoolingReleaseTemperatureOriginal(currentRawValue);
     this.log('Luxtronik cooling release test original value saved', {
       parameterIndex: parameter.index,
       rawValue: currentRawValue,
@@ -597,7 +605,72 @@ class LuxtronikDevice extends Device {
     return this.writeParameter(parameter.index, currentRawValue);
   }
 
-  async restoreCoolingReleaseTemperature() {
+  async saveCoolingReleaseTemperatureOriginal(currentRawValue) {
+    const savedRawValue = this.coolingReleaseTemperatureTestOriginalRaw
+      ?? this.getStoreValue(COOLING_RELEASE_TEST_ORIGINAL_STORE_KEY);
+
+    if (savedRawValue !== undefined && savedRawValue !== null) {
+      validateTemperatureRawValue(
+        VERIFIED_LUXTRONIK_PARAMETERS.COOLING_RELEASE_TEMPERATURE,
+        savedRawValue,
+      );
+      this.coolingReleaseTemperatureTestOriginalRaw = savedRawValue;
+      return savedRawValue;
+    }
+
+    await this.setStoreValue(COOLING_RELEASE_TEST_ORIGINAL_STORE_KEY, currentRawValue);
+    this.coolingReleaseTemperatureTestOriginalRaw = currentRawValue;
+    return currentRawValue;
+  }
+
+  queueCoolingReleaseOperation(operation) {
+    const previousOperation = this.coolingReleaseOperationQueue || Promise.resolve();
+    const currentOperation = previousOperation.catch(() => {}).then(operation);
+
+    this.coolingReleaseOperationQueue = currentOperation.catch(() => {});
+    return currentOperation;
+  }
+
+  setCoolingReleaseTemperature(value) {
+    return this.queueCoolingReleaseOperation(() => this.setCoolingReleaseTemperatureQueued(value));
+  }
+
+  async setCoolingReleaseTemperatureQueued(value) {
+    const parameter = VERIFIED_LUXTRONIK_PARAMETERS.COOLING_RELEASE_TEMPERATURE;
+    const temperature = Number(value);
+
+    if (!Number.isFinite(temperature)) {
+      throw new Error(`Cooling release temperature must be a number, received '${value}'.`);
+    }
+    if (!Number.isInteger(temperature * 2)) {
+      throw new Error(`Cooling release temperature ${temperature} °C must use 0.5 °C increments.`);
+    }
+
+    const rawValue = temperature * 10;
+    validateTemperatureRawValue(parameter, rawValue);
+
+    const parameters = await this.readParameters();
+    const currentRawValue = parameters[parameter.index];
+    validateTemperatureRawValue(parameter, currentRawValue);
+    const originalRawValue = await this.saveCoolingReleaseTemperatureOriginal(currentRawValue);
+
+    this.log('Luxtronik set cooling release temperature requested', {
+      parameterIndex: parameter.index,
+      luxtronikName: parameter.luxtronikName,
+      originalRawValue,
+      currentRawValue,
+      rawValue,
+      temperature,
+    });
+
+    return this.writeParameter(parameter.index, rawValue);
+  }
+
+  restoreCoolingReleaseTemperature() {
+    return this.queueCoolingReleaseOperation(() => this.restoreCoolingReleaseTemperatureQueued());
+  }
+
+  async restoreCoolingReleaseTemperatureQueued() {
     const parameter = VERIFIED_LUXTRONIK_PARAMETERS.COOLING_RELEASE_TEMPERATURE;
     const originalRawValue = this.coolingReleaseTemperatureTestOriginalRaw
       ?? this.getStoreValue(COOLING_RELEASE_TEST_ORIGINAL_STORE_KEY);
@@ -865,19 +938,22 @@ class LuxtronikDevice extends Device {
                 // for (const [i, value] of array_parameter.entries()) {
                 //   this.log(i, value);
                 // }
-                this.log("PARAM HEAT_ENERGY_INPUT" + array_parameter[1136]);
-                this.log("PARAM WATER_ENERGY_INPUT" + array_parameter[1137]);
-                this.log("PARAM POOL_ENERGY_INPUT" + array_parameter[1138]);
-                this.log("PARAM COOL_ENERGY_INPUT" + array_parameter[1139]);
-                this.log("PARAM SECOND_ENERGY_INPUT" + array_parameter[1140]);
-                let energyInputTotal = array_parameter[1136] + array_parameter[1137] + array_parameter[1138] + array_parameter[1139] + array_parameter[1140]
-                this.log("PARAM TOTAL_ENERGY_INPUT" + energyInputTotal);
-
-                this.energyInputHeat = (array_parameter[1136]);
-                this.energyInputCool = (array_parameter[1139]);
-                this.energyInputWater = (array_parameter[1137]);
-                this.energyInputPool = (array_parameter[1138]);
-                this.energyInputTotal = (array_parameter[1136] + array_parameter[1137] + array_parameter[1138] + array_parameter[1139] + array_parameter[1140]);
+                const energyInputs = readEnergyInputs(array_parameter);
+                if (energyInputs === null) {
+                  this.energyInputHeat = null;
+                  this.energyInputCool = null;
+                  this.energyInputWater = null;
+                  this.energyInputPool = null;
+                  this.energyInputTotal = null;
+                  this.log('Energy input parameters 1136-1140 are unavailable on this firmware');
+                } else {
+                  this.energyInputHeat = energyInputs.heat;
+                  this.energyInputCool = energyInputs.cool;
+                  this.energyInputWater = energyInputs.water;
+                  this.energyInputPool = energyInputs.pool;
+                  this.energyInputTotal = energyInputs.total;
+                  this.log('Luxtronik energy input values', energyInputs);
+                }
 
                 this.parametersArray = array_parameter;
 
@@ -904,7 +980,7 @@ class LuxtronikDevice extends Device {
                 this.energyTotal = (array_calculated[154]);
 
 
-                this.energyCurrent = (array_calculated[257]);
+                this.energyCurrent = readCurrentPower(array_calculated);
 
                 this.temperatureOutdoor = (array_calculated[15]);
                 this.temperatureHotGas = (array_calculated[14]);